Renamed variables related to step actions to be more descriptive. Split StepAction(s) and StepActionProvider into separate files
