+++ lib/code_guide_component.dart
@@ -2,7 +2,7 @@ import 'package:angular2/core.dart' show Component;
 
 import 'code_explanation_component.dart';
 import 'code_viewer_component.dart';
-import 'package:code_steps/step_cmds_directive.dart';
+import 'package:code_steps/step_action.dart';
 
 @Component(
   selector: 'code-guide',
