+++ lib/progression_service_provider.dart
@@ -1,11 +1,11 @@
 import 'package:angular2/core.dart';
 import 'lesson_loader.dart';
 import 'package:code_steps/progression_service.dart';
-import 'package:code_steps/step_cmds_directive.dart';
+import 'package:code_steps/step_actions_provider.dart';
 
 @Injectable()
-progressionServiceFactory(LessonLoader lessonLoader, StepActionsProvider stepCommandsDirective) =>
-    new ProgressionService(lessonLoader, stepCommandsDirective);
+progressionServiceFactory(LessonLoader lessonLoader, StepActionsProvider stepActionsProvider) =>
+    new ProgressionService(lessonLoader, stepActionsProvider);
 
 const progressionServiceProvider = const Provider(ProgressionService,
     useFactory: progressionServiceFactory,
