+++ web/main.dart
@@ -6,7 +6,7 @@ import 'package:angular2/platform/browser.dart';
 import 'package:code_steps/app_component.dart';
 import 'package:code_steps/lesson_loader.dart';
 import 'package:code_steps/progression_service_provider.dart';
-import 'package:code_steps/step_cmds_directive.dart';
+import 'package:code_steps/step_action.dart';
 
 main() {
   bootstrap(AppComponent, [LessonLoader, StepActionsProvider, progressionServiceProvider]);
